RT #118488: Proper normalization for tests in cpan/, dist
There is a function that is intended to transform from the "stripped"
test name (without leading ../ or trailing .t) to a usable relative
path. That function prepends '../t/' to tests, unless their name begins
with ext/ or lib/, in which case it prepends only '../'. This patch
adds cpan/ and dist/ to the list of special cases.