Create 'perl-current' directory and repository anew each run.
For the git synchronization problem described in:
https://rt.cpan.org/Ticket/Display.html?id=120008.

First, get the repository in 'git-perl' in sync with origin.  If smoking a
branch other than blead, remove any local branch by that name and create the
branch anew from origin so that we're getting the branch author's most recent
revisions.

Then, doing all that synching all over again in 'perl-current', simply remove
the perl-current directory entirely, and clone the now-up-to-date repo from
'git-perl' into it.  Then, proceed with smoke-testing as always.