Fix get_avail_w32compilers() by Steve Hay:
Test-Smoke-1.42 fails to detect the ccversion for my MSVC++
installations, which all reside under "C:\Program Files\...". It tries
to run a command like

"$map{ $CC }->{ccbin}" --version 2>&1

but the 'ccbin' there comes from whereis() and is already double-quoted
if it contains spaces, which causes an error like

'""C:\Program' is not recognized as an internal or external command,
operable program or batch file.

The attached patch fixes this.

https://rt.cpan.org/Ticket/Display.html?id=54542