Report all failing unit tests and sequences thereof.
The 'log' and 'out' files were correctly identifying failed unit tests and
sequences thereof, e.g.:

    ../lib/locale.t.............................................FAILED
        436-437, 441, 444, 458

However, only the first of these was being captured by
Test::Smoke::Reporter::_parse() -- called from within
Test::Smoke::Reporter::read_parse() -- and, hence, only the first was
available for recording in the 'jsn' file:

    "failures" : [
       {
          "status" : "FAILED",
          "extra" : [
             "436-437"
          ],
          "test" : "../lib/locale.t"
       }
    ],

And since it is the 'jsn' file that is sent to the CoreSmoke database -- from
which the web display draws its data -- the web page was only showing this:

    Test failures:
    ~~ ../lib/locale.t ................................ FAILED 436-437

Cf. http://perl5.test-smoke.org/report/68034

This commit improves the regex capturing the numbers of failing unit tests and
sequences of such tests.  In t/reporter.t one block of tests similar to
existing blocks is added.

The commit also adds to .gitignore one file created during testing.
For: RTC 125932