RT#132642: newer HTTP::Daemon broke my workaround
I used a workaround that replaced the hostname-part in $daemon->url with
the localhost IPv4 address (127.0.0.1).
From HTTP::Daemon 6.05 they started using IO::Socket::IP that treats the
IPv6 differently and on some platforms exclusively (so IPv4 doesn't work
on that socket).
From the [comments
in](https://github.com/libwww-perl/HTTP-Daemon/pull/40) I took the test
for `$daemon->sockhost` to determine IPv4 or IPv6, and now
replace the host part of the url with the correct address for localhost.
In the hope that HTTP::Daemon 6.08 will give a usable `$daemon->url`