Turn on autoflush for STDERR
On Windows there *appears* to be some buffering for STDERR.
I noticed this when troubleshooting/testing for issue #54: at some
point the STDERR output got split and was mixed between (later)
STDOUT output. [There is a sample log file in the ticket.]

What appears to fix it is to turn on autoflush for STDERR. Then the
logfile contains the expected output (when `print STDERR $err` is used).

So just to be avoid the issue in the future: do an explicit autoflush
of STDERR in the scripts.

(I do have some changes in mind where there could be output to both
 STDOUT and STDERR)