Abort tests when chdir to repo dir failed
The test init their own git repo and then does a `chdir()` to that
directory.

If that `chdir()` would fail then abort the test since it would/could otherwise
be modifying the Test-Smoke repo.

(I haven't seen failures with `chdir()`, it's more a safe-guard to prevent
 against altering the wrong repo)

(Using `die` instead of `skip` at the request of @abeltje)