Remove test repository when rsync isn't found
In t/syncer_rsync.t a fixed test directory (t/tsgit) is used.
At the end of the test it normally removes this directory but when
some of the tests are skipped then it left this directory in place
which caused the next test run to have a failure (on `git commit`).

Alternative fixes:
- use a tmp dir like t/syncer_git.t does
- remove the directory at the start of the test if it exist