Filter out the default msg in `user_skiped_tests`
`bin/configsmoke.pl` adds a default message when creating the 'skip tests' file.

Filter out that message in the report to reduce some 'useless' info.

For reference: the 'skip tests' file is used by `...::Smoker::set_skip_tests`
which does some more filtering (skip comments, ...).

Exact Filtering that it does:

	    $raw =~ m/^\s*#/ and next;
	    $raw =~ s/(\S+).*/$1/s;
	    if ($raw !~ m/\.t$/ and $raw !~ m/test\.pl$/) {
	        next;
	    }

I opted not to do the same filtering and instead just filter the default
message. If there are comments in the file then they could be useful
to include in the report (i.e. if they explain why a test file is skipped)