Fix `Test::Smoke::Util::set_local_patch`
The `set_local_patch` function got 'broken' when the 'patchlevel.h'
in the perl5 repo got detabified (committed on 17 Jan 2021).

It no longer worked because the code was using: `/\t,NULL/`
That worked fine when the tabs were used for indentation but it
got silently broken when the tabs were replaces with spaces.