Add USE_MINGW_ANSI_STDIO as default for Configure_win32()
    https://github.com/Perl/perl5/issues/20214

I guess this is a case of "Doctor, it hurts when I do this...". We'll
just switch it on (it won't hurt MSVC and benefits gcc) and provide and
escape with `-UUSE_MINGW_ANSI_STDIO`. We should aim to build the best
possible Perl on Windows that we can imo.