Make Test::Smoke installable as regular Perl module
In the past, most users of the software wanted the software installed in
a separate directory, but we never created an easy way to install as a
regular Perl module. This change makes it posible to do so:

	$ SMOKE_INST_SITE_LIB=1 perl Makefile.PL
or
	$ perl Makefile.PL --site-lib

Installing as a regular Perl module makes the scripts and documentation
more readily available.